Add some functionality
Add some initial functionality that we want, including:
- inserting a heading with title "Tasks"
- inserting a TODO subheading and some initial boilerplate text
- scheduling the TODO item for the current day
- setting a deadline on the TODO item for the next day

Add a function that will build up the filename based on the date (i.e. the base
directory followed by "YYYY/MM/DD.org") since it's a pretty logical way to keep
track of DMOs.

Add a base directory - I'll need to make sure it will be a configurable value -
in which DMO files will be created.
- need to figure out how to create intermediate directories so that we actually
do everything necessary for creating the file.